schoolName delete, blank modify
