+++ src/components/SuperAdmin/email/SuperAdminEmailContent.tsx
@@ -1,4 +1,4 @@
-import { User } from '@@types/index';
+import { Email } from '@@types/index';
 import { SubContainer } from '@components/common/container/AppContainer';
 import styled from '@emotion/styled';
 import { Color } from '@resources/colors';
@@ -9,31 +9,28 @@ const SubLabelContainer = styled.div`
   ${rowFlex()}
 `;
 
-const WorkspaceLabel = styled.div`
+const EmailLabel = styled.div`
   text-align: center;
   font-size: 18px;
   font-weight: 400;
   text-decoration: none;
-  color: #5c5c5c;
+  color: ${Color.GREY};
   cursor: pointer;
   transition: ease-in 0.1s;
   &:hover {
-    color: ${Color.KIO_ORANGE};
-    text-decoration: underline;
+    color: ${Color.KIO_ORANGE};ㄴ
   }
 `;
 
-function SuperAdminUserContent({ name, email, createdAt }: User) {
-  const datePart = createdAt.split('T')[0];
-  const filteredCreatedDate = datePart.replace(/-/g, '.');
-  const createdDateAndOwnerText = `${filteredCreatedDate} | ${email}`;
-
+function SuperAdminEmailContent({ name, domain }: Email) {
   return (
     <SubContainer useFlex={colFlex({ justify: 'center', align: 'start' })} customWidth={'1000px'} customHeight={'80px'} customGap={'5px'}>
-      <WorkspaceLabel className={'user-label'}>{name}</WorkspaceLabel>
-      <SubLabelContainer className={'sub-label-container'}>{createdDateAndOwnerText}</SubLabelContainer>
+      <EmailLabel className={'email-label'}>{name}</EmailLabel>
+      <SubLabelContainer className={'sub-label-container'}>
+        {name} | {domain}
+      </SubLabelContainer>
     </SubContainer>
   );
 }
 
-export default SuperAdminUserContent;
+export default SuperAdminEmailContent;
