+++ src/components/admin/order/NotPaidOrderCard.tsx
@@ -16,7 +16,7 @@ const Container = styled.div`
   padding: 22px;
   width: 350px;
   box-sizing: border-box;
-  background: ${Color.LIGHT_GREY};
+  background: #f7f7f7;
   gap: 7px;
   ${colFlex({ align: 'center' })}
 `;
