+++ src/components/admin/order/ToggleOrderCard.tsx
@@ -4,7 +4,6 @@ import styled from '@emotion/styled';
 import AppLabel from '@components/common/label/AppLabel';
 import ArrowUpSvg from '@resources/svg/ArrowUpSvg';
 import { colFlex, rowFlex } from '@styles/flexStyles';
-import { Color } from '@resources/colors';
 
 interface ToggleOrderCardProps {
   order: Order;
@@ -12,7 +11,7 @@ interface ToggleOrderCardProps {
 
 const Container = styled.div`
   width: 100%;
-  background: ${Color.LIGHT_GREY};
+  background: #f7f7f7;
   padding: 20px 50px;
   box-sizing: border-box;
   ${rowFlex({ justify: 'space-between' })}
