+++ src/hooks/admin/useAdminWorkspace.tsx
@@ -1,13 +1,12 @@
 import useApi from '@hooks/useApi';
 import { useSetRecoilState } from 'recoil';
-import { adminWorkspaceAtom } from '@recoils/atoms';
+import { userWorkspaceAtom } from '@recoils/atoms';
 import { Workspace } from '@@types/index';
 import { useNavigate } from 'react-router-dom';
 
 function useAdminWorkspace() {
   const { adminApi } = useApi();
-  const setAdminWorkspace = useSetRecoilState(adminWorkspaceAtom);
-
+  const setUserWorkspace = useSetRecoilState(userWorkspaceAtom);
   const navigate = useNavigate();
 
   const fetchWorkspace = (workspaceId: string | undefined | null) => {
@@ -16,7 +15,7 @@ function useAdminWorkspace() {
     adminApi
       .get<Workspace>('/workspace', { params: { workspaceId } })
       .then((res) => {
-        setAdminWorkspace(res.data);
+        setUserWorkspace(res.data);
       })
       .catch((error) => {
         alert(error.response.data.message);
@@ -24,18 +23,7 @@ function useAdminWorkspace() {
       });
   };
 
-  const updateWorkspaceTableCount = (workspaceId: string | undefined | null, tableCount: number) => {
-    adminApi
-      .post<Workspace>('/workspace/table-count', { workspaceId, tableCount })
-      .then((res) => {
-        setAdminWorkspace(res.data);
-      })
-      .catch((error) => {
-        alert(error.response.data.message);
-      });
-  };
-
-  return { fetchWorkspace, updateWorkspaceTableCount };
+  return { fetchWorkspace };
 }
 
 export default useAdminWorkspace;
