+++ src/pages/Admin/AdminWorkspace.tsx
@@ -1,14 +1,15 @@
 import React, { useEffect } from 'react';
-import { useNavigate, useParams } from 'react-router-dom';
+import { useParams } from 'react-router-dom';
 import useCustomNavigate from '@hooks/useCustomNavigate';
 import useAdminWorkspace from '@hooks/admin/useAdminWorkspace';
 import { useRecoilValue } from 'recoil';
-import { adminWorkspaceAtom } from '@recoils/atoms';
+import { userWorkspaceAtom } from '@recoils/atoms';
 import styled from '@emotion/styled';
 import ImageRouteButton from '@components/common/button/ImageRouteButton';
 import AppContainer from '@components/common/container/AppContainer';
 import AppFooter from '@components/common/footer/AppFooter';
 import orderImage from '@resources/image/orderImage.png';
+import orderHistoryImage from '@resources/image/orderHistoryImage.png';
 import productImage from '@resources/image/productImage.png';
 import orderManageImage from '@resources/image/orderManageImage.png';
 import { colFlex, rowFlex } from '@styles/flexStyles';
@@ -26,8 +27,7 @@ export const ButtonContainer = styled.div`
 function AdminWorkspace() {
   const { workspaceId } = useParams<{ workspaceId: string }>();
   const { fetchWorkspace } = useAdminWorkspace();
-  const navigate = useNavigate();
-  const workspace = useRecoilValue(adminWorkspaceAtom);
+  const workspace = useRecoilValue(userWorkspaceAtom);
   const { appendPath } = useCustomNavigate();
 
   useEffect(() => {
@@ -35,13 +35,11 @@ function AdminWorkspace() {
   }, []);
 
   return (
-    <AppContainer
-      useFlex={colFlex({ justify: 'center' })}
-      titleNavBarProps={{ title: workspace.name, subTitle: workspace.description, onLeftArrowClick: () => navigate('/admin') }}
-    >
+    <AppContainer useFlex={colFlex({ justify: 'center' })} titleNavBarProps={{ title: workspace.name, subTitle: workspace.description }}>
       <Container className={'admin-workspace-container'}>
         <ButtonContainer className={'button-container'}>
-          <ImageRouteButton src={orderImage} onClick={() => appendPath('/order')} buttonText={'주문 조회'} />
+          <ImageRouteButton src={orderImage} onClick={() => appendPath('/orders')} buttonText={'실시간 주문 조회'} />
+          <ImageRouteButton src={orderHistoryImage} onClick={() => appendPath('/orders-history')} buttonText={'전체 주문 조회'} />
           <ImageRouteButton src={productImage} onClick={() => appendPath('/products')} buttonText={'상품 관리'} />
           <ImageRouteButton src={orderManageImage} onClick={() => appendPath('/orders-manage')} buttonText={'주문 페이지 관리'} />
         </ButtonContainer>
