+++ src/pages/SuperAdmin/SuperAdminSearchContents.tsx
@@ -1,6 +1,5 @@
 import { PaginationResponse } from '@@types/index';
 import styled from '@emotion/styled';
-import { Color } from '@resources/colors';
 import React from 'react';
 
 const HorizontalLine = styled.hr`
@@ -10,7 +9,7 @@ const HorizontalLine = styled.hr`
 
 const EmptyLabel = styled.div`
   font-size: 40px;
-  color: ${Color.LIGHT_GREY};
+  color: #d8d8d8;
 `;
 
 interface ContentsProps {
