+++ src/pages/SuperAdmin/SuperAdminUser.tsx
@@ -4,12 +4,12 @@ import { userPaginationResponseAtom } from '@recoils/atoms';
 import { colFlex } from '@styles/flexStyles';
 import { useEffect, useRef } from 'react';
 import { useRecoilValue } from 'recoil';
+import SuperAdminSearchContents from './SuperAdminSearchContents';
 import Pagination from '@components/common/pagination/Pagination';
 import AppContainer from '@components/common/container/AppContainer';
 import useSuperAdminUser from '@hooks/SuperAdmin/useSuperAdminUser';
 import SuperAdminUserContent from '@components/SuperAdmin/user/SuperAdminUserContent';
 import { useNavigate } from 'react-router-dom';
-import SuperAdminSearchContents from '@components/SuperAdmin/SuperAdminSearchContents';
 
 const ContentContainer = styled.div<{ justifyCenter?: boolean }>`
   height: 550px;
